+++ core/makedeps.lua
@@ -25,11 +25,11 @@ local makeDeps = {
   write = function (self)
     self:add_modules()
     if type(self.filename) ~= "string" then
-      self.filename = SILE.masterFilename .. ".d"
+      self.filename = pl.path.splitext(SILE.input.filenames[1]) .. ".d"
     end
     local depfile, err = io.open(self.filename, "w")
     if not depfile then return SU.error(err) end
-    depfile:write(SILE.outputFilename .. ": " .. tostring(self._deps) .. "\n")
+    depfile:write(SILE.outputter:getOutputFilename() .. ": " .. tostring(self._deps) .. "\n")
     depfile:close()
   end
 }