+++ packages/color-fonts/init.lua
@@ -99,7 +99,8 @@ end
 
 package.documentation = [[
 \begin{document}
-The \autodoc:package{color-fonts} package adds support for fonts with a \code{COLR} OpenType table.
+The \autodoc:package{color-fonts} package adds support for fonts with multi-colored glyphs (that is,
+OpenType fonts with \code{COLR} and \code{CPAL} tables).
 This package is automatically loaded when such a font is detected.
 \end{document}
 ]]